% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recanaq-package.R
\docType{package}
\name{recanaq}
\alias{recanaq-package}
\alias{recanaq}
\title{Environment Canterbury Air Quality Data Package}
\description{
This package provides functions to retrieve air quality data from the
Environment Canterbury (ECan) data portal. It includes functions to:
}
\details{
\itemize{
\item List all monitoring stations (\code{get_stations()})
\item Retrieve daily measurements for a single station (\code{get_daily_one_station()})
\item Retrieve daily measurements for all stations (\code{get_daily_all_stations()})
}

The package handles date validation, tidy data formatting (long format),
and returns data as tibbles suitable for analysis with tidyverse tools.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bws200.github.io/recanaq}
  \item \url{https://github.com/bws200/recanaq}
}

}
\author{
\strong{Maintainer}: Ben Scott \email{ben.scott@ecan.govt.nz}

Authors:
\itemize{
  \item Ben Scott \email{ben.scott@ecan.govt.nz}
}

}
\keyword{internal}
